requirements-base.txt
starlette
uvicorn
graphene
fastapi
azure-servicebus